Move 'write_tmpl_file' from HydroTrend class to plugins.base module
All components (not just HydroTrend) need a Dakota tmpl file.
