+++ csdms/dakota/plugins/base.py
@@ -4,6 +4,59 @@
 from abc import ABCMeta, abstractmethod
 
 
+def write_tmpl_file(base_tmpl_file, base_input_file, parameter_names):
+    """Create a template HydroTrend input file used by Dakota.
+
+    The tmpl file is a HydroTrend input file, but with the
+    parameters used by Dakota replaced with their descriptors set
+    in the Dakota input file. The tmpl file is written to the
+    current directory.
+
+    Parameters
+    ----------
+    base_tmpl_file : str
+      The path to the template file defined for the HydroTrend
+      component.
+    base_input_file : str
+      A HydroTrend input file that contains parameter values for a
+      HydroTrend simulation.
+    parameter_names : list of str
+      A list of HydroTrend parameter names (standard names) that
+      will be evaluated by Dakota.
+
+    Returns
+    -------
+    str or None
+      The path to the new tmpl file, or None on an error.
+
+    """
+    import re
+
+    with open(base_tmpl_file, 'r') as fp:
+        txt_base_tmpl = fp.read().split('\n')
+    with open(base_input_file, 'r') as fp:
+        txt_base_input = fp.read().split('\n')
+
+    if len(txt_base_tmpl) != len(txt_base_input):
+        raise
+
+    for p_name in parameter_names:
+        for i, line_tmpl in enumerate(txt_base_tmpl):
+            if re.search(p_name, line_tmpl):
+                line_input = txt_base_input[i]
+                line_tmpl_split = line_tmpl.strip().split()
+                line_input_split = line_input.strip().split()
+                for j, item in enumerate(line_tmpl_split):
+                    if item.startswith('{' + p_name):
+                        line_input_split[j] = '{' + p_name + '}'
+                txt_base_input[i] = ' '.join(line_input_split)
+
+    tmpl_file = 'HYDRO.IN.tmpl'
+    with open(tmpl_file, 'w') as fp:
+        fp.write('\n'.join(txt_base_input))
+        return tmpl_file
+
+
 class PluginBase(object):
 
     """Describe features common to all Dakota plugins."""
