+++ csdms/dakota/tests/test_plugin_hydrotrend.py
@@ -14,6 +14,7 @@ import tempfile
 import numpy as np
 from numpy.testing import assert_almost_equal
 from nose.tools import raises, with_setup, assert_is, assert_true
+from csdms.dakota.plugins.base import write_tmpl_file
 from csdms.dakota.plugins.hydrotrend import HydroTrend, is_installed
 from csdms.dakota.utils import get_configuration
 from . import start_dir, data_dir
@@ -126,8 +127,8 @@ def test_write_tmpl_file():
     base_input_file = os.path.join(data_dir, 'HYDRO.IN.defaults')
     parameter_names = ['starting_mean_annual_temperature',
                        'total_annual_precipitation']
-    tmpl_file = HydroTrend.write_tmpl_file(base_tmpl_file,
-                                           base_input_file,
-                                           parameter_names)
+    tmpl_file = write_tmpl_file(base_tmpl_file,
+                                base_input_file,
+                                parameter_names)
     assert_true(filecmp.cmp(known_tmpl_file, tmpl_file))
     os.remove(tmpl_file)
